SDK Update
- This update will deprecate the following versions due to changes to the underlying SDK structure. **If you are playing on one of the versions listed below YOU SHOULD NOT UPDATE!**
    - ETS2 / ATS version 1.54.
    - ETS2 version 1.55 Beta.

**Changes**
- Now sending double the traffic information. `20 -> 40`
- Now sending double the light / gate information. `20 -> 40`
- Now sending 20% more route information. `5000 -> 6000 nodes`
- These improvements come after discovering how little performance impact the SDK has on the game. It makes sense to send more information if the performance is available.

This update *does not* include TMP data. That will happen later after I get the time. When doing these changes I realized how crap my virtual memory implementation is and it will be completely rewritten. The data is in the SDK, it just has to be sent to the backend.